curated_fastmcp_sqlite_teaching_case/requirements.txt
pandas>=2.0
matplotlib>=3.7
fastmcp>=2.12
python-docx>=1.2
python-pptx>=1.0
